Split up shared::contracts
