+++ crates/shared/src/lib.rs
@@ -7,7 +7,6 @@ pub mod bad_token;
 pub mod baseline_solver;
 pub mod code_fetching;
 pub mod code_simulation;
-pub mod contracts;
 pub mod conversions;
 pub mod current_block;
 pub mod db_order_conversions;
